Support out-of-core and distributed IVF_PQ ingestion (#531)

+++ src/include/detail/linalg/vector.h
@@ -194,7 +194,7 @@ void debug_vector(
       break;
     }
     std::cout << i;
-    if (i != end - 1) {
+    if (static_cast<size_t>(i) != end - 1) {
       std::cout << ", ";
     }
   }
