+++ src/include/test/unit_ivf_flat_group.cc
@@ -352,7 +352,7 @@ TEST_CASE(
   CHECK(x.get_previous_ingestion_timestamp() == expected_ingestion + offset);
   CHECK(x.get_previous_base_size() == expected_base + offset);
   CHECK(x.get_previous_num_partitions() == expected_partitions + offset);
-  CHECK(x.get_temp_size() == expected_temp_size + offset);
+  CHECK(x.get_temp_size() == expected_temp_size + static_cast<int64_t>(offset));
   CHECK(x.get_dimensions() == expected_dimension + offset);
 }
 
@@ -404,7 +404,7 @@ TEST_CASE("storage version", "[ivf_flat_group]") {
   CHECK(x.get_previous_ingestion_timestamp() == expected_ingestion + offset);
   CHECK(x.get_previous_base_size() == expected_base + offset);
   CHECK(x.get_previous_num_partitions() == expected_partitions + offset);
-  CHECK(x.get_temp_size() == expected_temp_size + offset);
+  CHECK(x.get_temp_size() == expected_temp_size + static_cast<int64_t>(offset));
   CHECK(x.get_dimensions() == expected_dimension + offset);
 }
